+++ Meraki.Api/MerakiClient.cs
@@ -292,7 +292,6 @@ public partial class MerakiClient : IDisposable
 			Mtu = RefitFor(Switch.Mtu),
 			LinkAggregations = RefitFor(Switch.LinkAggregations),
 			Ports = RefitFor(Switch.Ports),
-			PortsProfiles = RefitFor(Switch.PortsProfiles),
 			PortSchedules = RefitFor(Switch.PortSchedules),
 			QosRules = RefitFor(Switch.QosRules),
 			Routing = new()
@@ -330,8 +329,7 @@ public partial class MerakiClient : IDisposable
 			Devices = new()
 			{
 				ConnectionStats = RefitFor(Wireless.Devices.ConnectionStats),
-				LatencyStats = RefitFor(Wireless.Devices.LatencyStats),
-				OrganizationEthernetStatuses = RefitFor(Wireless.Devices.OrganizationEthernetStatuses)
+				LatencyStats = RefitFor(Wireless.Devices.LatencyStats)
 			},
 			ConnectionStats = RefitFor(Wireless.ConnectionStats),
 			FailedConnections = RefitFor(Wireless.FailedConnections),