+++ Meraki.Api/Sections/General/Devices/DevicesSection.cs
@@ -6,7 +6,7 @@ public partial class DevicesSection
 	internal IDevices Devices { get; set; } = null!;
 
 	public IDeviceClients Clients { get; internal set; } = null!;
-	public IDeviceCellularSims CellularSims { get; internal set } = null!;
+	public IDeviceCellularSims CellularSims { get; internal set; } = null!;
 	public IDeviceLldpCdp LldpCdp { get; internal set; } = null!;
 	public IDeviceLossAndLatencyHistory LossAndLatencyHistory { get; internal set; } = null!;
 	public IDeviceManagementInterface ManagementInterface { get; internal set; } = null!;