+++ Meraki.Api/Sections/Products/Wireless/WirelessDevicesSection.cs
@@ -4,5 +4,4 @@ public class WirelessDevicesSection
 {
 	public IWirelessDeviceConnectionStats ConnectionStats { get; internal set; } = null!;
 	public IWirelessDeviceLatencyStats LatencyStats { get; internal set; } = null!;
-	public IWirelessDeviceEthernetStatuses OrganizationEthernetStatuses { get; internal set; } = null!;
 }